This branch aims to replace Instaloader
The main action of this commit was the inclusion of InstaLogin.py
which at this current moment stores the username and passwords of
the logins inside of an encrypted json file. The encryption isn't
exactly the most well hidden but it does its job for now. In the
future I may try to hide the encryption key better if I feel the
need arises.

Other small details that were changed:
- Removed all mentions of Instaloader in Followertracker.py(Currently not functioning)
- Added if __name__=="__main__": to the end to make this the main script
- Gave the README.md a work in progress area for this branch
- Removed Herobrine